+++ src/commonMain/kotlin/csense/kotlin/tests/assertions/String.kt
@@ -7,7 +7,7 @@ import kotlin.test.assertNotEquals
 import kotlin.test.assertTrue
 
 /**
- * Asserts that the this string contains the given value.
+ * Asserts that this string contains the given value.
  * @receiver [String] the string to assert contains the given [value]
  * @param value [String] the value that should be in this string
  * @param ignoreCase [Boolean] if true, will ignore the casing, false means case sensitive.