+++ changelog.md
@@ -1,3 +1,12 @@
+# 0.0.53
+- kotlin 1.6.10
+- added 
+  - assertIsApply
+  - assertSizeAnd
+  - added assertSingle & assertEmpty for primitive array types
+  - added some @{int, etc}limit annotations 
+  - {Collection, Array}.assertNotEmpty
+
 # 0.0.52
 - kotlin 1.6.0